requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9    # Драйвер для PostgreSQL
SQLAlchemy==2.0.23        # ORM для удобной работы с БД
python-dotenv==1.0.0      # Для загрузки переменных окружения (.env)